move move button to wishlist view screen
