+++ client/src/components/WishlistItem.svelte
@@ -2,11 +2,14 @@
     import '../tailwind.css';
 
     import { createEventDispatcher } from 'svelte';
-    import { Patch } from '../http';
+    import { Get, Patch } from '../http';
     import { makeRoutes } from '../routes';
+    import Modal from './Modal.svelte';
+    import RadioList from './RadioList.svelte';
 
     export let wishlist_item: IWishlistItem;
     export let is_edit: boolean = false;
+    export let is_owned: boolean = false;
     export let share_guid: string = null;
     export let has_other_wishlists: boolean = false;
 
@@ -19,11 +22,25 @@
     let link: string = wishlist_item.link;
     let notes: string = wishlist_item.notes;
     let bought: boolean = wishlist_item.bought;
+    let move_item_modal: Modal;
+    let target_wishlist: IWishlist = null;
 
     $: {
         wishlist_item.link = link;
         wishlist_item.notes = notes;
         wishlist_item.bought = bought;
+    }
+
+    async function get_all_wishlists(): Promise<IWishlist[]> {
+        const payload = await Get<IWishlist[]>(Api.Wishlists.GetAllForUser);
+
+        let wishlists = payload.get_json();
+
+        if (wishlist_item.wishlist_id != null) {
+            wishlists = wishlists.filter(w => w.id != wishlist_item.wishlist_id);
+        }
+
+        return wishlists;
     }
 
     async function mark_as_bought() {
@@ -34,6 +51,26 @@
         }
 
         dispatch('bought');
+    }
+
+    async function move_item_to_list() {
+        const confirmed = await move_item_modal.show();
+
+        if (!confirmed) {
+            return;
+        }
+
+        await Patch(Api.WishlistItems.PatchReparent.append(wishlist_item.id).append(target_wishlist.id), {});
+
+        dispatch('moved', wishlist_item);
+
+        // const item_index = wishlist_items.indexOf(item);
+        // if (item_index === -1) {
+        //     return;
+        // }
+
+        // wishlist_items.splice(item_index, 1);
+        // wishlist_items = wishlist_items; // trigger reactivity
     }
 </script>
 
@@ -47,11 +84,7 @@
             </div>
         </div>
         <div class="">
-            {#if has_other_wishlists}
-            <button class="icon-button" on:click={() => dispatch('move', wishlist_item)}>
-                <span class="fa-solid fa-arrow-right-from-bracket"></span>
-            </button>
-            {/if}
+
             <button class="icon-button" on:click={() => dispatch('delete', wishlist_item)}>
                 <span class="fa-solid fa-trash pointer-events-none"></span>
             </button>
@@ -66,6 +99,11 @@
             </div>
         </div>
         <div class="text-base text-slate-600">
+            {#if has_other_wishlists && is_owned}
+            <button class="icon-button" on:click={() => move_item_to_list()}>
+                <span class="fa-solid fa-arrow-right-from-bracket"></span>
+            </button>
+            {/if}
             <button class="icon-button" on:click={mark_as_bought}>
                 <span class="fa-solid fa-basket-shopping pointer-events-none"></span>
             </button>
@@ -73,6 +111,24 @@
     </div>
     {/if}
 </div>
+
+<Modal bind:this={move_item_modal} id="move">
+    <span slot="header">
+        Move item
+    </span>
+    <span slot="body">
+        {#await get_all_wishlists()}
+            Loading...
+        {:then wishlists} 
+            <p>Choose the wishlist to move this item to:</p>
+            <RadioList group_id="wishlists_group" items={wishlists} bind:selected_item={target_wishlist}></RadioList>
+        {/await}
+    </span>
+    <span slot="buttons" let:close_modal={close}>
+        <button class="button" on:click={() => close()}>Cancel</button>
+        <button class="button" on:click={() => close("true")}>Move item</button>
+    </span>
+</Modal>
 
 <style lang="less">
 </style>
